add v1.1 and change rev to ver
