--- keyboards/otaku_split/rev1.1/keymaps/takashiski/readme.md
@@ -1,1 +1,0 @@
-# The default keymap for otaku_split
