+++ keyboards/otaku_split/ver0/info.json
