+++ keyboards/otaku_split/ver1.1/keymaps/bootmagic/config.h
