+++ keyboards/otaku_split/ver1.1/keymaps/bootmagic/readme.md
